v0.0.26
New combat coming along nicely.
Player can move, then the monster now moves on its own.
There is an "attack" button for the player, however it is only testing
to see if you can or cannot attack.
Added a logging text box on the combat page.
Next Task is going to be displaying Monster/Player information.
Fixed issue with the combat field being resized weirdly.
Fixed issue where monster is awkwardly stuck in a corner. Now it knows
what to do when it is trapped in the corner...